Moved api2.js into index.js
